Add id exception handling

+++ src/main/java/com/codecool/submarinemanager/crewman/CrewmanServiceImpl.java
@@ -2,6 +2,7 @@ package com.codecool.submarinemanager.crewman;
 
 import com.codecool.submarinemanager.crewman.Crewman;
 import com.codecool.submarinemanager.crewman.CrewmanRepository;
+import com.codecool.submarinemanager.exception.IdDoesNotExistException;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
@@ -18,15 +19,28 @@ public class CrewmanServiceImpl implements CrewmanService {
         return crewmanRepository.findAll();
     }
 
-    public Crewman findCrewmanById(Integer id) {
-        return crewmanRepository.findOne(id);
+    public Crewman findCrewmanById(Integer id) throws IdDoesNotExistException {
+
+        Crewman crewman = crewmanRepository.findOne(id);
+
+        if (crewman == null) {
+            throw new IdDoesNotExistException();
+        }
+
+        return crewman;
     }
 
     public void saveCrewman(Crewman crewman) {
         crewmanRepository.save(crewman);
     }
 
-    public void deleteCrewman(Integer id) {
+    public void deleteCrewman(Integer id) throws IdDoesNotExistException {
+        Crewman crewman = crewmanRepository.findOne(id);
+
+        if (crewman == null) {
+            throw new IdDoesNotExistException();
+        }
+
         crewmanRepository.delete(id);
     }
 }
